+++ packages/bot-utils/src/utils/array/sort/order-asc-number.ts
@@ -1,0 +1,3 @@
+export function orderAscNumber<T>(list: T[], propertyKey: keyof T): T[] {
+  return list.sort((prev, cur) => Number(prev[propertyKey]) - Number(cur[propertyKey]));
+}