+++ packages/bot-utils/src/utils/index.ts
@@ -1,3 +1,3 @@
 export * from './recursive-promise-callback';
 export * from './sleep';
-export * from './unique';
+export * from './array';